Use `github.com/spf13/cobra` module to generate command line interface
The "command subcommand" command line approach will facilitate the expansion of the tool's command line interface as
additional functionality is added. This is provided Cobra, which is the established standard for Arduino tooling Go
applications.

The previous interface:

```
libraries-repository-engine [CONFIG_FILE [REGISTRY_FILE]]
```

The new interface:

```
libraries-repository-engine sync [--config-file=CONFIG_FILE] [REGISTRY_FILE]
```

Backwards compatibility with the old interface has been preserved, but it is deprecated. A warning will be displayed when
it is used.
