+++ go.mod
@@ -7,7 +7,8 @@ require (
 	github.com/arduino/go-paths-helper v1.6.1
 	github.com/arduino/golang-concurrent-workers v0.0.0-20170202182617-6710cdc954bc
 	github.com/go-git/go-git/v5 v5.4.2
-	github.com/google/go-cmp v0.5.2 // indirect
+	github.com/spf13/cobra v1.2.1
+	github.com/spf13/pflag v1.0.5
 	github.com/stretchr/testify v1.7.0
 	github.com/vaughan0/go-ini v0.0.0-20130923145212-a98ad7ee00ec
 	go.bug.st/relaxed-semver v0.0.0-20190922224835-391e10178d18
